requirements.txt
langchain_community
langchain_openai
langchain
langchain_core
langchain_experimental
langchainhub
python-dotenv
faiss-cpu
pypdf==4.3.0
langchain-ai21

streamlit==1.34.0
